# frontend/docs/contratoAPIswagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Relatorio de Faturamento API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
        "/recursos/{id}": {
            "get": {
                "tags": [
                    "Recursos"
                ],
                "summary": "Buscar recurso por ID",
                "description": "Busca um recurso pelo ID.",
                "operationId": "buscarPorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do recurso",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RecursoResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Recursos"
                ],
                "summary": "Atualizar recurso",
                "description": "Atualiza um recurso existente.",
                "operationId": "atualizar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do recurso",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecursoRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Recurso atualizado com sucesso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RecursoResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    },
                    "404": {
                        "description": "Recurso não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Recursos"
                ],
                "summary": "Deletar recurso",
                "description": "Remove um recurso do sistema.",
                "operationId": "deletar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do recurso",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Recurso removido com sucesso"
                    },
                    "404": {
                        "description": "Recurso não encontrado"
                    }
                }
            }
        },
        "/ordens-servico/{id}": {
            "get": {
                "tags": [
                    "Ordens de Serviço"
                ],
                "summary": "Buscar ordem de serviço por ID",
                "description": "Busca uma ordem de serviço pelo ID.",
                "operationId": "buscarPorId_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da ordem de serviço",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ordem de serviço encontrada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrdemServicoResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ordem de serviço não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Ordens de Serviço"
                ],
                "summary": "Atualizar ordem de serviço",
                "description": "Atualiza uma ordem de serviço existente.",
                "operationId": "atualizar_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da ordem de serviço",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrdemServicoRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "404": {
                        "description": "Ordem de serviço não encontrada"
                    },
                    "400": {
                        "description": "Dados inválidos"
                    },
                    "200": {
                        "description": "Ordem de serviço atualizada com sucesso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrdemServicoResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ordens de Serviço"
                ],
                "summary": "Deletar ordem de serviço",
                "description": "Remove uma ordem de serviço do sistema.",
                "operationId": "deletar_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da ordem de serviço",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Ordem de serviço removida com sucesso"
                    },
                    "404": {
                        "description": "Ordem de serviço não encontrada"
                    }
                }
            }
        },
        "/feriados/{id}": {
            "get": {
                "tags": [
                    "Feriados"
                ],
                "summary": "Buscar feriado por ID",
                "description": "Busca um feriado pelo ID.",
                "operationId": "buscarPorId_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do feriado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Feriado não encontrado"
                    },
                    "200": {
                        "description": "Feriado encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FeriadoResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Feriados"
                ],
                "summary": "Atualizar feriado",
                "description": "Atualiza um feriado existente.",
                "operationId": "atualizar_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do feriado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FeriadoRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Feriado atualizado com sucesso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FeriadoResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    },
                    "404": {
                        "description": "Feriado não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Feriados"
                ],
                "summary": "Deletar feriado",
                "description": "Remove um feriado do sistema.",
                "operationId": "deletar_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do feriado",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Feriado não encontrado"
                    },
                    "204": {
                        "description": "Feriado removido com sucesso"
                    }
                }
            }
        },
        "/faturamentos/{id}": {
            "get": {
                "tags": [
                    "Faturamentos"
                ],
                "summary": "Buscar faturamento por ID",
                "description": "Busca um faturamento pelo ID.",
                "operationId": "buscarPorId_3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do faturamento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Faturamento encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FaturamentoResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Faturamento não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Faturamentos"
                ],
                "summary": "Atualizar faturamento",
                "description": "Atualiza um faturamento existente.",
                "operationId": "atualizar_3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do faturamento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FaturamentoRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Faturamento atualizado com sucesso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FaturamentoResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    },
                    "404": {
                        "description": "Faturamento não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Faturamentos"
                ],
                "summary": "Deletar faturamento",
                "description": "Remove um faturamento do sistema.",
                "operationId": "deletar_3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do faturamento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Faturamento removido com sucesso"
                    },
                    "404": {
                        "description": "Faturamento não encontrado"
                    }
                }
            }
        },
        "/faturamento-detalhes/{id}": {
            "get": {
                "tags": [
                    "Faturamento Detalhe"
                ],
                "summary": "Buscar detalhe de faturamento por ID",
                "description": "Busca um detalhe de faturamento pelo ID.",
                "operationId": "buscarPorId_4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do detalhe de faturamento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhe encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FaturamentoDetalheResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Detalhe não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Faturamento Detalhe"
                ],
                "summary": "Atualizar detalhe de faturamento",
                "description": "Atualiza um detalhe de faturamento existente.",
                "operationId": "atualizar_4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do detalhe de faturamento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FaturamentoDetalheRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "404": {
                        "description": "Detalhe não encontrado"
                    },
                    "400": {
                        "description": "Dados inválidos"
                    },
                    "200": {
                        "description": "Detalhe atualizado com sucesso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FaturamentoDetalheResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Faturamento Detalhe"
                ],
                "summary": "Deletar detalhe de faturamento",
                "description": "Remove um detalhe de faturamento do sistema.",
                "operationId": "deletar_4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do detalhe de faturamento",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Detalhe não encontrado"
                    },
                    "204": {
                        "description": "Detalhe removido com sucesso"
                    }
                }
            }
        },
        "/contratos/{codigoContrato}": {
            "get": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Buscar contrato por código",
                "description": "Busca um contrato pelo código.",
                "operationId": "buscarPorId_5",
                "parameters": [
                    {
                        "name": "codigoContrato",
                        "in": "path",
                        "description": "Código do contrato",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contrato encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContratoResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Contrato não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Atualizar contrato",
                "description": "Atualiza um contrato existente.",
                "operationId": "atualizar_5",
                "parameters": [
                    {
                        "name": "codigoContrato",
                        "in": "path",
                        "description": "Código do contrato",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContratoRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "404": {
                        "description": "Contrato não encontrado"
                    },
                    "200": {
                        "description": "Contrato atualizado com sucesso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ContratoResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contratos"
                ],
                "summary": "Deletar contrato",
                "description": "Remove um contrato do sistema.",
                "operationId": "deletar_5",
                "parameters": [
                    {
                        "name": "codigoContrato",
                        "in": "path",
                        "description": "Código do contrato",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Contrato removido com sucesso"
                    },
                    "404": {
                        "description": "Contrato não encontrado"
                    }
                }
            }
        },
...
        "/feriados": {
            "get": {
                "tags": [
                    "Feriados"
                ],
                "summary": "Listar feriados",
                "description": "Lista todos os feriados cadastrados.",
                "operationId": "listarTodos_3",
                "responses": {
                    "200": {
                        "description": "Lista de feriados retornada com sucesso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FeriadoResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Feriados"
                ],
                "summary": "Criar feriado",
                "description": "Cria um novo feriado no sistema.",
                "operationId": "criar_2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FeriadoRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Dados inválidos"
                    },
                    "201": {
                        "description": "Feriado criado com sucesso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FeriadoResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "RecursoRequest": {
                "type": "object",
                "properties": {
                    "codigoContrato": {
                        "type": "string"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "fatorAjuste": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "description": "Dados do recurso"
            },
            "RecursoResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "codigoContrato": {
                        "type": "string"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "fatorAjuste": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "OrdemServicoRequest": {
                "type": "object",
                "properties": {
                    "codigoContrato": {
                        "type": "string"
                    },
                    "numeroOs": {
                        "type": "string"
                    }
                },
                "description": "Dados da ordem de serviço"
            },
            "OrdemServicoResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "codigoContrato": {
                        "type": "string"
                    },
                    "numeroOs": {
                        "type": "string"
                    }
                }
            },
            "FeriadoRequest": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "description": "Dados do feriado"
            },
            "FeriadoResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "data": {
                        "type": "string",
                        "format": "date"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}